// src/types.ts
export interface Address { street: string; suite: string; city: string; zipcode: string; }
export interface Company { name: string }
export interface User {
  id: number;
  name: string;
  email: string;
  phone?: string;
  website?: string;
  address?: Address;
  company?: Company;
}
export type SortKey = 'name' | 'email';
export type SortOrder = 'asc' | 'desc';